+++ src/AssimpWorker/import/AMLImporter/CMakeLists.txt
@@ -8,17 +8,11 @@
 SET(IMPORT_AML_HPP_FILES
 	${CMAKE_CURRENT_SOURCE_DIR}/AMLImporter.hpp
 	${CMAKE_CURRENT_SOURCE_DIR}/AMLFrameImporter.hpp
-	${CMAKE_CURRENT_SOURCE_DIR}/ColladaMassager.hpp
-	${CMAKE_CURRENT_SOURCE_DIR}/ColladaRecursiveImporter.hpp
-	${CMAKE_CURRENT_SOURCE_DIR}/ColladaMassagerRegistry.hpp
 )
 
 SET(IMPORT_AML_SRC_FILES
 	${CMAKE_CURRENT_SOURCE_DIR}/AMLImporter.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/AMLFrameImporter.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/ColladaMassager.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/ColladaRecursiveImporter.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/ColladaMassagerRegistry.cpp
 )
 
 set(import_AML_HPP_FILES ${IMPORT_AML_HPP_FILES} PARENT_SCOPE)
