+++ src/AssimpWorker/import/ColladaImporter/CMakeLists.txt
@@ -7,10 +7,16 @@
 
 SET(IMPORT_COLLADA_HPP_FILES
 	${CMAKE_CURRENT_SOURCE_DIR}/ColladaImporter.hpp
+	${CMAKE_CURRENT_SOURCE_DIR}/ColladaMassager.hpp
+	${CMAKE_CURRENT_SOURCE_DIR}/ColladaRecursiveImporter.hpp
+	${CMAKE_CURRENT_SOURCE_DIR}/ColladaMassagerRegistry.hpp
 )
 
 SET(IMPORT_COLLADA_SRC_FILES
 	${CMAKE_CURRENT_SOURCE_DIR}/ColladaImporter.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/ColladaMassager.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/ColladaRecursiveImporter.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/ColladaMassagerRegistry.cpp
 )
 
 set(import_COLLADA_HPP_FILES ${IMPORT_COLLADA_HPP_FILES} PARENT_SCOPE)
