+++ src/AssimpWorker/import/ColladaImporter/ColladaMassager.cpp
