+++ src/AssimpWorker/import/ColladaImporter/ColladaMassagerRegistry.cpp
