+++ src/AssimpWorker/import/ColladaImporter/ColladaMassagerRegistry.hpp
