+++ src/AssimpWorker/import/ColladaImporter/ColladaRecursiveImporter.cpp
